+++ src/Services/StudyAreaService.js
@@ -234,6 +234,9 @@ var StreamStats;
                         _this.toaster.clear();
                         _this.eventManager.RaiseEvent(Services.onSelectedStudyAreaChanged, _this, StudyAreaEventArgs.Empty);
                         _this.canUpdate = true;
+                        if (_this.regionService.selectedRegion.Applications.indexOf("HydrologicFeatures") != -1) {
+                            _this.queryHydrologicFeatures();
+                        }
                     }
                     else {
                         _this.clearStudyArea();
@@ -547,51 +550,95 @@ var StreamStats;
                     _this.toaster.pop('error', "There was an HTTP error querying coordinated reach", "Please retry", 0);
                 });
             };
-            StudyAreaService.prototype.queryHydrologyAttributes = function () {
+            StudyAreaService.prototype.queryHydrologicFeatures = function () {
                 var _this = this;
-                var ppt = this.snappedPourPoint;
-                var turfPoint = turf.point([ppt[0], ppt[1]]);
-                var distance = 0.05;
-                var bearings = [-90, 0, 90, 180];
-                var boundingBox = [];
-                bearings.forEach(function (bearing, index) {
-                    var destination = turf.destination(turfPoint, distance, bearing);
-                    boundingBox[index] = destination.geometry.coordinates[index % 2 == 0 ? 0 : 1];
+                var snappedDelineationPoint = turf.point([this.snappedPourPoint[0], this.snappedPourPoint[1]]);
+                var delineatedBasinGeometry = this.selectedStudyArea.FeatureCollection.features[1].geometry;
+                var NHDStreamIntersections = {};
+                var nhdLayerOptions = {
+                    "url": configuration.baseurls['NationalMapServices'] + configuration.queryparams['NHDQueryService']
+                };
+                var nhdLayer = L.esri.dynamicMapLayer(nhdLayerOptions);
+                nhdLayer.query().intersects(delineatedBasinGeometry).where("GNIS_ID IS NOT NULL")
+                    .run(function (error, results) {
+                    if (error) {
+                        _this.toaster.pop('error', "There was an error querying NHD streams", error, 0);
+                        _this.selectedStudyArea.NHDStreamIntersections = [];
+                    }
+                    else if (results && results.features.length > 0) {
+                        results.features.forEach(function (feature) {
+                            if (feature.properties.GNIS_ID) {
+                                var stream = turf.lineString(feature.geometry.coordinates);
+                                var distanceToPourPoint = turf.pointToLineDistance(snappedDelineationPoint, stream) * 1000 * 3.28084;
+                                if (Object.keys(NHDStreamIntersections).indexOf(feature.properties.GNIS_ID) == -1) {
+                                    NHDStreamIntersections[feature.properties.GNIS_ID] = { "GNIS_ID": feature.properties.GNIS_ID, "GNIS_NAME": feature.properties.GNIS_NAME, "distanceToPourPoint": distanceToPourPoint };
+                                }
+                                else {
+                                    if (distanceToPourPoint < NHDStreamIntersections[feature.properties.GNIS_ID]["distanceToPourPoint"]) {
+                                        NHDStreamIntersections[feature.properties.GNIS_ID]["distanceToPourPoint"] = distanceToPourPoint;
+                                    }
+                                }
+                            }
+                        });
+                        _this.selectedStudyArea.NHDStreamIntersections = Object.keys(NHDStreamIntersections).map(function (key) { return NHDStreamIntersections[key]; });
+                        var minDistanceToPourPointFeature = _this.selectedStudyArea.NHDStreamIntersections.reduce(function (prev, curr) {
+                            return prev.distanceToPourPoint < curr.distanceToPourPoint ? prev : curr;
+                        });
+                        _this.selectedStudyArea.NHDStream = minDistanceToPourPointFeature;
+                        _this.selectedStudyArea.Disclaimers['hydrologicFeatures'] = true;
+                    }
+                    else if (results && results.features.length == 0) {
+                        _this.selectedStudyArea.NHDStreamIntersections = [];
+                    }
+                    else {
+                        _this.toaster.pop('error', "There was an error querying NHD streams", "Please retry", 0);
+                        _this.selectedStudyArea.NHDStreamIntersections = [];
+                    }
                 });
-                var outFieldsNHD = "GNIS_ID,GNIS_NAME";
-                var urlNHD = configuration.baseurls['NationalMapServices'] + configuration.queryparams['NHDQueryService']
-                    .format(this.selectedStudyArea.RegionID.toLowerCase(), boundingBox[0], boundingBox[1], boundingBox[2], boundingBox[3], this.selectedStudyArea.Pourpoint.crs, outFieldsNHD);
-                var requestNHD = new WiM.Services.Helpers.RequestInfo(urlNHD, true);
-                var self = this;
-                this.Execute(requestNHD).then(function (response) {
-                    if (response.data.error) {
-                        _this.toaster.pop('error', "There was an error querying NHD stream lines", response.data.error.message, 0);
-                        return;
+                var WBDHUC8Intersections = {};
+                var wbdLayerOptions = {
+                    "url": configuration.baseurls['NationalMapServices'] + configuration.queryparams['WBDQueryService']
+                };
+                var wbdLayer = L.esri.dynamicMapLayer(wbdLayerOptions);
+                wbdLayer.query().intersects(delineatedBasinGeometry).where("huc8 IS NOT NULL")
+                    .run(function (error, results) {
+                    if (error) {
+                        _this.toaster.pop('error', "There was an error querying WBD HUC 8 watersheds", error, 0);
+                        _this.selectedStudyArea.WBDHUC8Intersections = [];
                     }
-                    if (response.data.features.length > 0) {
-                        var attributes = response.data.features[0].attributes;
-                        console.log(attributes);
-                        self.selectedStudyArea.NHDStream = attributes;
+                    else if (results && results.features.length > 0) {
+                        results.features.forEach(function (feature) {
+                            if (feature.properties.huc8) {
+                                var huc8 = turf.polygon(feature.geometry.coordinates);
+                                if (turf.booleanPointInPolygon(snappedDelineationPoint, huc8)) {
+                                    var distanceToPourPoint = 0;
+                                }
+                                else {
+                                    var distanceToPourPoint = turf.pointToLineDistance(snappedDelineationPoint, turf.polygonToLineString(huc8)) * 1000 * 3.28084;
+                                }
+                                if (Object.keys(WBDHUC8Intersections).indexOf(feature.properties.huc8) == -1) {
+                                    WBDHUC8Intersections[feature.properties.huc8] = { "huc8": feature.properties.huc8, "name": feature.properties.name, "distanceToPourPoint": distanceToPourPoint };
+                                }
+                                else {
+                                    if (distanceToPourPoint < WBDHUC8Intersections[feature.properties.huc8]["distanceToPourPoint"]) {
+                                        WBDHUC8Intersections[feature.properties.huc8]["distanceToPourPoint"] = distanceToPourPoint;
+                                    }
+                                }
+                            }
+                        });
+                        _this.selectedStudyArea.WBDHUC8Intersections = Object.keys(WBDHUC8Intersections).map(function (key) { return WBDHUC8Intersections[key]; });
+                        var minDistanceToPourPointFeature = _this.selectedStudyArea.WBDHUC8Intersections.reduce(function (prev, curr) {
+                            return prev.distanceToPourPoint < curr.distanceToPourPoint ? prev : curr;
+                        });
+                        _this.selectedStudyArea.WBDHUC8 = minDistanceToPourPointFeature;
                     }
-                }, function (error) {
-                    _this.toaster.pop('error', "There was an error querying NHD stream lines", "Please retry", 0);
-                });
-                var outFieldsWBD = "huc8,name";
-                var urlWBD = configuration.baseurls['NationalMapServices'] + configuration.queryparams['WBDQueryService']
-                    .format(this.selectedStudyArea.RegionID.toLowerCase(), boundingBox[0], boundingBox[1], boundingBox[2], boundingBox[3], this.selectedStudyArea.Pourpoint.crs, outFieldsWBD);
-                var requestWBD = new WiM.Services.Helpers.RequestInfo(urlWBD, true);
-                this.Execute(requestWBD).then(function (response) {
-                    if (response.data.error) {
-                        _this.toaster.pop('error', "There was an error querying WBD HUC8 watersheds", response.data.error.message, 0);
-                        return;
+                    else if (results && results.features.length == 0) {
+                        _this.selectedStudyArea.WBDHUC8Intersections = [];
                     }
-                    if (response.data.features.length > 0) {
-                        var attributes = response.data.features[0].attributes;
-                        console.log(attributes);
-                        self.selectedStudyArea.WBDHUC8 = attributes;
+                    else {
+                        _this.toaster.pop('error', "There was an error querying WBD HUC 8 watersheds", "Please retry", 0);
+                        _this.selectedStudyArea.WBDHUC8Intersections = [];
                     }
-                }, function (error) {
-                    _this.toaster.pop('error', "There was an error querying WBD HUC8 watersheds", "Please retry", 0);
                 });
             };
             StudyAreaService.prototype.queryRegressionRegions = function () {