+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
     -  Added SC Storm Runoff model
     -  Added Urban Bohman Method functionality to the SC Storm Runoff model
-    -  South Carolina application returns NHD stream info and HUC8 info
+    -  South Carolina application selects nearest NHD stream and HUC 8 watershed and adds to report
+    -  South Carolina application allows user to select a different NHD stream and HUC 8 watershed
   ### Changed
 
   ### Fixed